build_utils: replace fypp with submodule
The advantage is that it is obvious which version of Fypp we use and it
can be easily updated. The disadvantage is that the release process gets
more complicated since we have to create a tarball which includes a
clone of the Fypp submodule, see:

  https://help.github.com/articles/creating-releases/
  https://stackoverflow.com/questions/34719785/how-to-add-submodule-files-to-a-github-release

On the other hand, this can possibly be automated using some CI system
whenever a new tag gets added.