travis: move codecov upload into MPI+OpenMP make stage
cmake seems to have a problem of correctly deducing the module
dependencies with pre-processor directives interrupting the `use`
statements, see

  https://gitlab.kitware.com/cmake/cmake/issues/18188

Without those line-number statements, the coverage report is unable to
match the proper lines, giving false measures.
So, revert to using the make-generated build for generating coverage
reports and wait for a cmake+ninja-based build.