libcusmm/predict: fix bug occuring when two simultaneous runs
In the case in which two training runs for the same algorithm are run
simultaneously, they may try to create the same logging folder, since
its path is based on the timestamp. In that case, the folder is created
by the first run, and the second run crashes as it tries to create a
folder which already exists. This commit fixes this issue by catching
the FileExistsError and creating a logging folder with a slightly
different name.