cmake: patch FindBLAS to check OpenBLAS for OMP
Uses the openblas_get_parallel() function to check whether OpenBLAS is
built with OMP. If not, looks for other OpenBLAS library candidates
libopenblas_omp.so/.a and libopenblas_openmp.so/.a.

FindLAPACK.cmake is needed because it calls FindBLAS, otherwise
it will use the CMake provided FindBLAS.cmake, not our patched one.

fixes #295