Only compile *.cu with device compiler
Currently all of acc is compiler with the device compiler, which happens
to work when gcc is the host compiler & openmp is disabled; or maybe if
clang is both host and device compiler and openmp is enabled; but _does
not_ work for gcc host, clang device and openmp enabled. This will cause
linking to different omp implementations.

The workaround is to remove HIP_SOURCE_PROPERTY_FORMAT 1 from acc
sources, so that only *.cu is compiled by the device compiler. But since
certain hip-specific headers are included, we have to add the
__HIP_PLATFORM_HCC__ define (or possibly __HIP_PLATFORM_NVCC__) for the
host compiler.

TODO: check if the host compiler can avoid ending up with those
hip-headers so that we don't have to manually add those defines.