Remove gpu arch passed to hiprtc
JIT compilation fails in DBCSR with newer rocm versions (>=4.1)
because there is a mismatch detected between the target id passed to
hiprtcCompileProgram and what hiprtc already passes by default.

hiprtcCompileProgram now honors compile options passed to it. It
also by default has always passed the target ids of the default
device to the compiler, so we don't have to pass this architecture
string as a compile option again. This also helps avoid any
changes to DBCSR when target ids are modified for future AMD GPUs.