Add support for mpi_f08
This PR adds support for the new mpi_f08 of the MPI 3.0 standard. This module solves several interface issues in the MPI backend (Type/Kind/Rank explosion, ASYNCHRONOUS, memory allocation, ...). Because not all MPI libraries in use support the mpi_f08 module properly, the use of the new interface is an opt-in solution, the old MPI interface remains the default until the support has improved (MPICH compiles but segfaults with certain versions of gfortran).